Refactor GUI app boundary and remove theme/tauri duplicate shell
