+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Ajisai - An AI-first, vector-oriented, fractional-dataflow programming language.">
-    <meta name="theme-color" content="#007bff">
+    <meta name="theme-color" content="#6b5b95">
 
 
     <link rel="manifest" href="./manifest.json">
@@ -15,22 +15,8 @@
 
     <title>Ajisai</title>
 
-
     <script src="./ajisai-config.js?v=20260410"></script>
-    <script src="./ajisai-theme.js?v=20260410"></script>
-    <style id="theme-vars"></style>
-    <script>
 
-        if (typeof AjisaiTheme !== 'undefined') {
-            const vars = AjisaiTheme.getAll();
-            let css = ':root {';
-            for (const [key, value] of Object.entries(vars)) {
-                css += `${key}: ${value};`;
-            }
-            css += '}';
-            document.getElementById('theme-vars').textContent = css;
-        }
-    </script>
 
     <link rel="stylesheet" href="public/ajisai-base.css?v=20260408">
     <link rel="stylesheet" href="app-interface.css?v=20260408">
