--- public/ajisai-theme.js
@@ -1,271 +1,0 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-const AjisaiTheme = {
-
-
-
-
-    hexToRgb: function(hex) {
-        const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-        return result ? {
-            r: parseInt(result[1], 16),
-            g: parseInt(result[2], 16),
-            b: parseInt(result[3], 16)
-        } : null;
-    },
-
-    rgbToHex: function(r, g, b) {
-        return '#' + [r, g, b].map(x => {
-            const hex = Math.max(0, Math.min(255, Math.round(x))).toString(16);
-            return hex.length === 1 ? '0' + hex : hex;
-        }).join('');
-    },
-
-    getLuminance: function(hex) {
-        const rgb = this.hexToRgb(hex);
-        if (!rgb) return 0.5;
-        const [r, g, b] = [rgb.r, rgb.g, rgb.b].map(v => {
-            v /= 255;
-            return v <= 0.03928 ? v / 12.92 : Math.pow((v + 0.055) / 1.055, 2.4);
-        });
-        return 0.2126 * r + 0.7152 * g + 0.0722 * b;
-    },
-
-    isLight: function(hex) {
-        return this.getLuminance(hex) > 0.5;
-    },
-
-    lighten: function(hex, amount) {
-        const rgb = this.hexToRgb(hex);
-        if (!rgb) return hex;
-        return this.rgbToHex(
-            rgb.r + (255 - rgb.r) * amount,
-            rgb.g + (255 - rgb.g) * amount,
-            rgb.b + (255 - rgb.b) * amount
-        );
-    },
-
-    darken: function(hex, amount) {
-        const rgb = this.hexToRgb(hex);
-        if (!rgb) return hex;
-        return this.rgbToHex(
-            rgb.r * (1 - amount),
-            rgb.g * (1 - amount),
-            rgb.b * (1 - amount)
-        );
-    },
-
-
-    getTextColor: function(bgHex) {
-        return this.isLight(bgHex) ? '#333333' : '#ffffff';
-    },
-
-
-    getTextColorSecondary: function(bgHex) {
-        return this.isLight(bgHex) ? '#555555' : '#cccccc';
-    },
-
-
-    getTextColorTertiary: function(bgHex) {
-        return this.isLight(bgHex) ? '#777777' : '#aaaaaa';
-    },
-
-
-    getBorderColor: function(bgHex) {
-
-        return this.isLight(bgHex) ? this.darken(bgHex, 0.25) : this.lighten(bgHex, 0.3);
-    },
-
-
-    getTintedDark: function(hex, baseValue = 0.12, tintAmount = 0.25) {
-        const rgb = this.hexToRgb(hex);
-        if (!rgb) return '#2d2d2d';
-        return this.rgbToHex(
-            baseValue * 255 + rgb.r * tintAmount,
-            baseValue * 255 + rgb.g * tintAmount,
-            baseValue * 255 + rgb.b * tintAmount
-        );
-    },
-
-
-
-
-
-    getPrimaryColor: function() {
-        if (typeof AjisaiConfig !== 'undefined' && AjisaiConfig.primaryColor) {
-            return AjisaiConfig.primaryColor;
-        }
-
-        console.warn('AjisaiConfig not found, using fallback color');
-        return '#6b5b95';
-    },
-
-
-
-
-
-    getBackgrounds: function() {
-        const p = this.getPrimaryColor();
-        return {
-
-            header: p,
-
-            body: this.lighten(p, 0.94),
-
-            article: this.lighten(p, 0.82),
-
-            section: this.lighten(p, 0.88),
-
-            menu: this.lighten(p, 0.12),
-
-            code: this.getTintedDark(p)
-        };
-    },
-
-
-
-
-
-    generateVariables: function() {
-        const primary = this.getPrimaryColor();
-        const bg = this.getBackgrounds();
-        const secondary = this.lighten(primary, 0.2);
-        const rgb = this.hexToRgb(primary);
-
-        return {
-
-            "--color-primary": primary,
-            "--color-secondary": secondary,
-
-
-            "--bg-header": bg.header,
-            "--bg-body": bg.body,
-            "--bg-article": bg.article,
-            "--bg-section": bg.section,
-            "--bg-menu": bg.menu,
-            "--bg-menu-hover": this.lighten(bg.menu, 0.15),
-            "--bg-code": bg.code,
-            "--bg-code-inline": `rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, 0.1)`,
-
-
-            "--text-on-header": this.getTextColor(bg.header),
-            "--text-on-header-secondary": this.getTextColorSecondary(bg.header),
-            "--text-on-header-tertiary": this.getTextColorTertiary(bg.header),
-            "--text-on-body": this.getTextColor(bg.body),
-            "--text-on-body-secondary": this.getTextColorSecondary(bg.body),
-            "--text-on-article": this.getTextColor(bg.article),
-            "--text-on-section": this.getTextColor(bg.section),
-            "--text-on-menu": this.getTextColor(bg.menu),
-            "--text-on-menu-hover": this.isLight(bg.menu) ? '#000000' : '#ffffff',
-            "--text-on-code": "#f8f8f2",
-
-
-            "--heading-h1": this.getTextColor(bg.header),
-            "--heading-h2": this.isLight(bg.article) ? this.darken(primary, 0.1) : this.lighten(primary, 0.3),
-            "--heading-h3": this.isLight(bg.article) ? primary : this.lighten(primary, 0.5),
-
-
-            "--text-code-comment": this.lighten(primary, 0.4),
-
-
-            "--border-on-header": this.getBorderColor(bg.header),
-            "--border-on-body": this.getBorderColor(bg.body),
-            "--border-on-article": this.getBorderColor(bg.article),
-            "--border-on-section": this.getBorderColor(bg.section),
-            "--border-on-menu": this.getBorderColor(bg.menu),
-            "--border-on-code": this.getBorderColor(bg.code),
-
-
-            "--radius-main": "10px",
-
-
-            "--font-stack-primary": '"Inter", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Hiragino Kaku Gothic ProN", Meiryo, sans-serif',
-            "--font-stack-mono": '"SF Mono", Consolas, "Liberation Mono", Monaco, monospace',
-
-
-            "--max-width-container": "1200px",
-            "--breakpoint-mobile": "768px",
-            "--col-left-flex": "1",
-            "--col-right-flex": "1",
-
-
-            "--border-color": this.getBorderColor(bg.article),
-            "--border-main": `solid 1px ${this.getBorderColor(bg.article)}`,
-            "--color-secondary": this.getBorderColor(bg.article),
-            "--gradient-header": bg.header,
-            "--gradient-background": bg.body,
-            "--gradient-parent": bg.article,
-            "--gradient-child": bg.section,
-            "--gradient-menu": bg.menu,
-            "--gradient-menu-hover": this.lighten(bg.menu, 0.15),
-            "--header-text": this.getTextColor(bg.header),
-            "--header-text-secondary": this.getTextColorSecondary(bg.header),
-            "--header-text-tertiary": this.getTextColorTertiary(bg.header),
-            "--menu-text": this.getTextColor(bg.menu),
-            "--menu-text-hover": this.isLight(bg.menu) ? '#000000' : '#ffffff',
-            "--color-light": bg.body,
-            "--color-medium": this.darken(bg.body, 0.1),
-            "--color-border-strong": this.darken(bg.body, 0.25),
-            "--color-dark": bg.article,
-            "--color-text": this.getTextColor(bg.body),
-            "--color-text-light": this.getTextColorSecondary(bg.body),
-            "--code-bg": bg.code,
-            "--code-text": "#f8f8f2",
-            "--code-comment": this.lighten(primary, 0.4),
-            "--code-inline-bg": `rgba(${rgb.r}, ${rgb.g}, ${rgb.b}, 0.1)`,
-
-
-            "--color-core": "#E65100",
-            "--color-dependency": "#E69500",
-            "--color-non-dependency": "#009B68",
-            "--color-stack": "#990099",
-            "--color-module": "#0277BD",
-
-
-            "--color-signature-map": "#DFEBFD",
-            "--color-signature-form": "#DDEFE3",
-            "--color-signature-fold": "#FCE3D8"
-        };
-    },
-
-
-
-
-
-    apply: function() {
-        const root = document.documentElement;
-        const vars = this.generateVariables();
-        for (const [key, value] of Object.entries(vars)) {
-            root.style.setProperty(key, value);
-        }
-    },
-
-    getAll: function() {
-        return this.generateVariables();
-    }
-};
-
-if (typeof window !== 'undefined') {
-    window.AjisaiTheme = AjisaiTheme;
-}
