+++ public/manifest.json
@@ -6,7 +6,7 @@
   "scope": "./",
   "display": "standalone",
   "background_color": "#f5f5f5",
-  "theme_color": "#007bff",
+  "theme_color": "#6b5b95",
   "orientation": "any",
   "icons": [
     {
